requirements.txt
mlflow-skinny==2.15.0
awswrangler==3.9.0
pandas==2.2.2
seaborn==0.13.2
scikit-learn==1.5.1
optuna==3.6.1
xgboost==2.1.1